+++ app/src/main/java/com/example/apiproject/Pokemon.kt
@@ -1,0 +1,9 @@
+package com.example.apiproject
+
+
+
+data class Pokemon(
+	val name: String,
+	val types: List<String>,
+	val imageUrl: String
+)
